# docs/openapi.json
{
  "swagger": "2.0",
  "basePath": "/",
  "paths": {
...
    "/v1/auth/forgot-password": {
      "post": {
        "responses": {
          "200": {
            "description": "If request accepted (email sent if user exists)"
          },
          "400": {
            "description": "Invalid email format"
          }
        },
        "summary": "Request password recovery",
        "description": "Sends email with reset link if the user exists.",
        "operationId": "auth_forgot_password",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ForgotPasswordRequest"
            }
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "responses": {
          "200": {
            "description": "Login successful - returns JWT token",
            "schema": {
              "$ref": "#/definitions/TokenResponse"
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many login attempts - rate limited"
          }
        },
        "summary": "Authenticate user and get JWT token",
        "description": "Use the returned token in the `Authorization` header for authenticated requests:\n`Authorization: Bearer <token>`",
        "operationId": "auth_login",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/rotas/": {
      "post": {
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/RotaResponse"
            }
          }
        },
        "operationId": "create_rota",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RotaCreateRequest"
            }
          }
        ],
        "tags": [
          "rotas"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RotaListResponse"
            }
          }
        },
        "operationId": "list_rotas",
        "tags": [
          "rotas"
        ]
      }
    },
...
    "/v1/rotas/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "put": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "update_rota",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RotaUpdateRequest"
            }
          }
        ],
        "tags": [
          "rotas"
        ]
      },
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "delete_rota",
        "tags": [
          "rotas"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RotaDetailResponse"
            }
          }
        },
        "operationId": "get_rota",
        "tags": [
          "rotas"
        ]
      }
    },
    "/v1/rotas/{id}/horarios": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "post": {
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/RotaHorarioResponse"
            }
          }
        },
        "operationId": "add_rota_horario",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RotaHorarioCreateRequest"
            }
          }
        ],
        "tags": [
          "rotas"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RotaHorarioListResponse"
            }
          }
        },
        "operationId": "list_rota_horarios",
        "tags": [
          "rotas"
        ]
      }
    },
    "/v1/rotas/{id}/inscricao": {
      "parameters": [
        {
          "in": "path",
          "description": "UUID da Rota",
          "name": "id",
          "required": true,
          "type": "string"
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "inscrever_aluno",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RotaInscricaoRequest"
            }
          }
        ],
        "tags": [
          "rotas"
        ]
      }
    },
    "/v1/rotas/{id}/pontos": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "add_rota_pontos",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RotaPontosAddRequest"
            }
          }
        ],
        "tags": [
          "rotas"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PontoFlatListResponse"
            }
          }
        },
        "operationId": "list_rota_pontos",
        "tags": [
          "rotas"
        ]
      }
    },
...
  },
  "info": {
    "title": "BusKá API",
    "version": "1.0.0",
    "description": "\n## Sistema de Gerenciamento de Transporte Escolar\n\nAPI para gerenciamento de rotas, viagens e alunos do transporte escolar municipal.\n\n### Autenticação\nTodos os endpoints (exceto `/auth/login`) requerem autenticação JWT.\nInclua o header: `Authorization: Bearer <seu_token>`\n\n### Roles\n- **ALUNO**: Visualiza rotas, confirma presença em viagens\n- **MOTORISTA**: Gerencia viagens atribuídas, inicia/finaliza trajetos\n- **GESTOR**: Acesso completo à prefeitura (CRUD de rotas, motoristas, relatórios)\n        "
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT token. Format: Bearer <token>"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "definitions": {
    "LoginRequest": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "description": "Email do usuário"
        },
        "password": {
          "type": "string",
          "description": "Senha"
        }
      },
      "type": "object"
    },
    "TokenResponse": {
      "properties": {
        "message": {
          "type": "string",
          "description": "Mensagem de sucesso"
        },
        "token": {
          "type": "string",
          "description": "JWT Token"
        },
        "user": {
          "description": "Dados do usuário",
          "allOf": [
            {
              "$ref": "#/definitions/UserInfo"
            }
          ]
        }
      },
      "type": "object"
    },
    "UserInfo": {
      "properties": {
        "id": {
          "type": "string",
          "description": "ID do usuário"
        },
        "nome": {
          "type": "string",
          "description": "Nome do usuário"
        },
        "email": {
          "type": "string",
          "description": "Email do usuário"
        },
        "role": {
          "type": "string",
          "description": "Perfil do usuário"
        }
      },
      "type": "object"
    },
    "ForgotPasswordRequest": {
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "description": "Email para recuperação de senha"
        }
      },
      "type": "object"
    },
...
    "RotaCreateRequest": {
      "required": [
        "nome"
      ],
      "properties": {
        "nome": {
          "type": "string",
          "description": "Nome da rota"
        },
        "motorista_padrao_id": {
          "type": "string",
          "description": "UUID do motorista padrão"
        },
        "veiculo_padrao_id": {
          "type": "string",
          "description": "UUID do veículo padrão"
        },
        "pontos": {
          "type": "array",
          "description": "Pontos da rota",
          "items": {
            "$ref": "#/definitions/RotaPontoAddRequest"
          }
        },
        "horarios": {
          "type": "array",
          "description": "Grade de horários",
          "items": {
            "$ref": "#/definitions/RotaHorarioCreateRequest"
          }
        }
      },
      "type": "object"
    },
    "RotaPontoAddRequest": {
      "required": [
        "ordem",
        "ponto_id"
      ],
      "properties": {
        "ponto_id": {
          "type": "string",
          "description": "UUID do ponto"
        },
        "ordem": {
          "type": "integer",
          "description": "Ordem na rota"
        }
      },
      "type": "object"
    },
    "RotaHorarioCreateRequest": {
      "required": [
        "dias",
        "horario_saida",
        "sentido"
      ],
      "properties": {
        "horario_saida": {
          "type": "string",
          "description": "Horário (HH:MM)"
        },
        "sentido": {
          "type": "string",
          "description": "IDA, VOLTA ou CIRCULAR"
        },
        "dias": {
          "type": "array",
          "description": "Dias da semana (SEG, TER, ...)",
          "items": {
            "type": "string"
          }
        }
      },
      "type": "object"
    },
    "RotaResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID"
        },
        "nome": {
          "type": "string",
          "description": "Nome"
        },
        "motorista_id": {
          "type": "string",
          "description": "UUID do motorista padrão"
        },
        "veiculo_id": {
          "type": "string",
          "description": "UUID do veículo padrão"
        },
        "prefeitura_id": {
          "type": "string",
          "description": "UUID da prefeitura"
        },
        "municipio_nome": {
          "type": "string",
          "description": "Nome do município"
        },
        "municipio_uf": {
          "type": "string",
          "description": "UF do município"
        }
      },
      "type": "object"
    },
    "RotaListResponse": {
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RotaResponse"
          }
        },
        "total": {
          "type": "integer",
          "description": "Total de rotas"
        }
      },
      "type": "object"
    },
    "RotaInscricaoRequest": {
      "required": [
        "acao"
      ],
      "properties": {
        "acao": {
          "type": "string",
          "description": "inscrever ou desinscrever"
        }
      },
      "type": "object"
    },
    "RotaPontosAddRequest": {
      "properties": {
        "pontos": {
          "type": "array",
          "description": "Pontos da rota",
          "items": {
            "$ref": "#/definitions/RotaPontoAddRequest"
          }
        }
      },
      "type": "object"
    },
    "PontoFlatListResponse": {
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PontoFlatResponse"
          }
        },
        "total": {
          "type": "integer",
          "description": "Total de pontos"
        }
      },
      "type": "object"
    },
    "PontoFlatResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID do ponto"
        },
        "apelido": {
          "type": "string",
          "description": "Nome/apelido"
        },
        "latitude": {
          "type": "number",
          "description": "Latitude"
        },
        "longitude": {
          "type": "number",
          "description": "Longitude"
        },
        "ordem": {
          "type": "integer",
          "description": "Ordem"
        }
      },
      "type": "object"
    },
    "RotaHorarioResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID"
        },
        "horario_saida": {
          "type": "string",
          "description": "Horário"
        },
        "sentido": {
          "type": "string",
          "description": "Sentido"
        },
        "dias": {
          "type": "array",
          "description": "Dias",
          "items": {
            "type": "string"
          }
        }
      },
      "type": "object"
    },
    "RotaHorarioListResponse": {
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RotaHorarioResponse"
          }
        },
        "total": {
          "type": "integer",
          "description": "Total de horários"
        }
      },
      "type": "object"
    },
    "RotaUpdateRequest": {
      "required": [
        "nome"
      ],
      "properties": {
        "nome": {
          "type": "string",
          "description": "Nome da rota"
        },
        "motorista_padrao_id": {
          "type": "string",
          "description": "UUID do motorista padrão"
        },
        "veiculo_padrao_id": {
          "type": "string",
          "description": "UUID do veículo padrão"
        },
        "pontos": {
          "type": "array",
          "description": "Pontos da rota",
          "items": {
            "$ref": "#/definitions/RotaPontoAddRequest"
          }
        },
        "horarios": {
          "type": "array",
          "description": "Grade de horários",
          "items": {
            "$ref": "#/definitions/RotaHorarioCreateRequest"
          }
        }
      },
      "type": "object"
    },
    "RotaDetailResponse": {
      "properties": {
        "id": {
          "type": "string",
          "description": "UUID"
        },
        "nome": {
          "type": "string",
          "description": "Nome"
        },
        "motorista_id": {
          "type": "string",
          "description": "UUID do motorista padrão"
        },
        "veiculo_id": {
          "type": "string",
          "description": "UUID do veículo padrão"
        },
        "prefeitura_id": {
          "type": "string",
          "description": "UUID da prefeitura"
        },
        "municipio_nome": {
          "type": "string",
          "description": "Nome do município"
        },
        "municipio_uf": {
          "type": "string",
          "description": "UF do município"
        },
        "pontos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PontoFlatResponse"
          }
        },
        "horarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RotaHorarioResponse"
          }
        }
      },
      "type": "object"
    },
...
  },
...
}